check headless and move to pipeline.py
